+++ R/processing_step.R
@@ -12,43 +12,63 @@ processing_step <- function(...)
 
 #' @rdname processing_step
 #' 
-#' @param fun_call A function call created with as.list(match.call())
+#' @param fun either a function or a non-empty character string naming the function to be called.
+#' @param arglist a list of arguments to the function call. The names attribute of args gives the argument names.
 #' @param type A string description of the step
-#' @param id A unique id associated with the step
+#' @param name A name associated with the step. Will be converted to a unique id.
 #' @param ... further arguments passed to or from other methods(not
 #'   currenctly used).
-#' @return An object of calss `processing_step`
+#' @return An object of class `processing_step`
 #' @export
 #' @examples
 #' library(tidySpectR)
 #' 
-#' do_something <- function(x, y, z){
-#'      print("doing something")
-#'      as.list(match.call())
-#' }
-#'
-#' fun_call <- do_something(1,2,3)
-#' # Remove unwanted arguments
-#' fun_call$x <- NULL
-#' fun_call
+#' # Creating a masking step
+#' stp <- 
+#'    processing_step(
+#'      fun = mask, 
+#'      arglist = list(from = 5, to = Inf),
+#'      type = "masking", 
+#'      name = "masking")
 #' 
-#' processing_step(fun_call, "test", "test_01")
+#' stp
 #' @importFrom rlang is_callable
-processing_step.default <- function(fun_call, type= "none", id = rand_id("none"), ...){
+processing_step.default <- function(fun, arglist, type= "none", name = "none", ...){
     
-    if (!is_callable(fun_call[[1]])){
-        rlang::abort("`fun_call` must be a function call.")
+    if (!is.character(fun)){
+        fun = as.character(substitute(fun))
     }
     
     process <- 
-        list(id = id,
+        list(id = rand_id(name),
              type = type,
-             fun = fun_call[[1]],
-             fun_call = fun_call,
-             arglist = fun_call[-1] )
+             fun = fun,
+             arglist = arglist)
     
     class(process) <- "processing_step"
     return(process)
+}
+
+#' Applies a `processing_step` on a `collection` object
+#' @rdname process
+#' @param x A `processing_step` object
+#' @param collection a `collection` to process
+#' @examples
+#' library(tidySpectR)
+#' 
+#' # Creating a masking step
+#' stp <- 
+#'    processing_step(
+#'      fun = mask, 
+#'      arglist = list(from = 5, to = Inf),
+#'      type = "masking", 
+#'      name = "masking")
+#' 
+#' process(stp, fa_nmr)
+#' @export
+process.processing_step <- function(x, collection, ...){
+    arglist <- append(list(collection), x$arglist)
+    do.call(x$fun, arglist)
 }
 
 #' Tidy a `processing_step`
@@ -61,17 +81,14 @@ processing_step.default <- function(fun_call, type= "none", id = rand_id("none")
 #' @examples
 #' library(tidySpectR)
 #' 
-#' do_something <- function(x, y, z){
-#'      print("doing something")
-#'      as.list(match.call())
-#' }
-#'
-#' fun_call <- do_something(1,2,3)
-#' # Remove unwanted arguments
-#' fun_call$x <- NULL
-#' fun_call
+#' # Creating a masking step
+#' stp <- 
+#'    processing_step(
+#'      fun = mask, 
+#'      arglist = list(from = 5, to = Inf),
+#'      type = "masking", 
+#'      name = "masking")
 #' 
-#' stp <- processing_step(fun_call, "test", "test_01")
 #' tidy(stp)
 #' @rdname tidy.processing_step
 #' @export
@@ -79,62 +96,20 @@ processing_step.default <- function(fun_call, type= "none", id = rand_id("none")
 #' @importFrom tidyr unnest_wider
 tidy.processing_step <- function(x, ...){
     tibble(type = x$type,
-           method = deparse(x$fun),
+           method = x$fun,
            args = list(x$arglist),
            id = x$id) %>% 
     unnest_wider(args)
 } 
 
+#' @export
+#' @importFrom rlang call2
 print.processing_step <- function(x, ...){
-    cat(
-        deparse(
-            as.call(x$fun_call))
-    )
-    cat("\n")
+    funcall <-
+        do.call(
+            rlang::call2, 
+            append(x$fun, x$arglist)
+        )
+    print(funcall)
     invisible(x)
-}
-
-#' Applies a processing step on a `collection` object
-#'
-#' @aliases process process.processing_step
-#' @export
-process <- function(x, ...)
-    UseMethod("process")
-
-#' @rdname process
-#' @param x A `processing_step` object to apply
-#' @param collection A `collection` object to process
-#' @param ... further arguments passed to or from other methods(not
-#'   currenctly used).
-#' @return The retrun value of the processing step call
-#' @export
-#' @examples
-#' library(tidySpectR)
-#' 
-#' # Creating a function call to `mask`
-#' # Note that the `collection` argument is omitted
-#' fun_call <- as.list(match.call(mask, call("mask", from = 5, to = Inf)))
-#' 
-#' stp <- processing_step(fun_call, "masking", "masking_01")
-#' processed <- process(stp, fa_nmr)
-#' processed
-process.processing_step <- function(x, collection, ...){
-    arglist <- append(list(collection), x$arglist)
-    do.call(eval(x$fun), arglist)
-}
-
-#' Generate random id
-#' @param id A base string to expand with random digits
-#' @ param ... currently unused
-#' @importFrom stringr str_pad
-#' @importFrom stats runif
-#' @keywords internal
-rand_id <- function(id, ...){
-    paste0(
-        id, 
-        "_", 
-        str_pad(
-            floor(
-                runif(1, min = 0, max = 999999)), 
-        6, pad = "0"))
 }